requirements.txt
pylint
flake8
django
djangorestframework
requests
celery
redis
django-celery-beat
